added additional fiber functions to orthotropic and two fiber family

+++ src/elements/solid/materials/J2plasticity/SSJ2LinHardBase2D.h
@@ -1,4 +1,4 @@
-/* $Id: SSJ2LinHardBase2D.h,v 1.3 2004-08-01 01:04:19 paklein Exp $ */
+/* $Id: SSJ2LinHardBase2D.h,v 1.4 2006-10-24 00:11:20 thao Exp $ */
 /* created: paklein (02/12/1997)                                          */
 /* Interface for a elastoplastic material that is linearly                */
 /* isotropically elastic subject to the Huber-von Mises yield             */
@@ -92,7 +92,7 @@ protected:
 	bool fplastic;
 
 	/* element level internal variables */
-	double fnstatev;             // number of internal state variables 
+	int fnstatev;             // number of internal state variables 
 	dArrayT fstatev;             // internal state variable array
 
 	/*current values*/
